refactor(*): rename component
