+++ example/src/app/medium/child.component.ts
